Add spawn island and void world generation Add item Astrolabe. For pack makers to make their own spawn islands It saves a 32x cube. I might make some kind of interaction to control the size Add config file Fix lang file. Still some stuff missing. I'll clean it up later Stop JEI from yelling at me Try to implement albedo lights and fail Add kindling block. You place it and rub a stick against it and it sets on fire eventually and burns and drops ash Add ash. It works exactly like bonemeal. I might change that Clean up JSON loading Make crucibles only insert one item

+++ src/main/java/com/jozufozu/exnihiloomnia/common/entity/EntityThrownStone.java
@@ -14,6 +14,7 @@ import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 
 public class EntityThrownStone extends EntityThrowable
@@ -53,7 +54,10 @@ public class EntityThrownStone extends EntityThrowable
                 IBlockState blockState = this.world.getBlockState(blockPos);
     
                 //We want it to break glass, but some stuff (barrels) shouldn't
-                if (blockState.getMaterial() == Material.GLASS && !(blockState.getBlock() instanceof ITileEntityProvider))
+    
+                Vec3d motion = new Vec3d(this.motionX, this.motionY, this.motionZ);
+                
+                if (motion.lengthSquared() >= 1.2 && blockState.getMaterial() == Material.GLASS && blockState.getBlockHardness(world, blockPos) <= 0.6f && !(blockState.getBlock() instanceof ITileEntityProvider))
                 {
                     this.world.destroyBlock(blockPos, false);
                     this.motionX *= 0.8f;
